Add kustomize manifests
Requests to this service will be proxied through an nginx proxy server
configured via the Google Service Management API. This is a requirement
when using Google Cloud Endpoints to correctly configure the service
according to the OpenAPI configuration file.

Additionally the manifests declare a single deployment (with 2 replicas)
running both the Extensible Service Proxy (nginx container mentioned
above) and the gateway container. The image for the gateway container
must be specified on the command line when deploying this service.