Describe the API surface through OpenAPI
Following a RESTful approach to API design describe the API surface of
the Gateway service using both OpenAPI to enforce the schema and
JSON:API to define the structure of the messages.

The Gateway service is ultimately responsible for orchestrating requests
to downstream services and in addition composing responses from these
services into a single message to be returned to the client. Therefore
this service defines endpoints that may result in multiple service being
called within a single request.