Add pact contract testing tool
Contract testing is a way to ensure that services (such as an API
provider and a client) can communicate with each other. Without contract
testing, the only way to know that services can communicate is by using
expensive and brittle integration tests.

This commit verifies the provider side of the contract, using the HTTP
interface exposed by the API in order to exercise the entire stack.

Note in the tests the gRPC upstream services have been mocked using
GoMock, providing test isolation yet ensuring the interface used whilst
testing is the same as production. This is important to avoid nasty
surprises when this is deployed in a live environment.